+++ LastMile/APIDecodableTypes/Int8+APIDecodable.swift
@@ -1,5 +1,5 @@
 //
-//  UInt64+Decodable.swift
+//  Int8+APIDecodable.swift
 //  LastMile
 //
 //  Created by Josh Elkins on 3/2/19.
@@ -9,4 +9,4 @@
 import Foundation
 
 
-extension UInt64: JSONRawValueType {}
+extension Int8: JSONRawValueType {}
