+++ LastMile/APIDecodableTypes/UInt16+APIDecodable.swift
@@ -1,5 +1,5 @@
 //
-//  UInt32+Decodable.swift
+//  UInt16+APIDecodable.swift
 //  LastMile
 //
 //  Created by Josh Elkins on 3/2/19.
@@ -9,4 +9,4 @@
 import Foundation
 
 
-extension UInt32: JSONRawValueType {}
+extension UInt16: JSONRawValueType {}
