+++ LastMile/APIDecodableTypes/UInt64+APIDecodable.swift
@@ -1,5 +1,5 @@
 //
-//  Float+Decodable.swift
+//  UInt64+APIDecodable.swift
 //  LastMile
 //
 //  Created by Josh Elkins on 3/2/19.
@@ -9,4 +9,4 @@
 import Foundation
 
 
-extension Float: JSONRawValueType {}
+extension UInt64: JSONRawValueType {}
